// dsl/packages/cli/test/test-specs/gen-out/client/typescript-client/src/model/ScalarRecord_Substitution.ts
// Generated by RSD - Do not modify
import { isRange, type Range, RangeFromJSON, RangeToJSON } from './_Scalars.js';
import { checkOptProp, checkProp, createTypedArrayGuard, isNull, isRecord, isString, isUndefined, propMappedListValue, propMappedValue } from '../_type-utils.js';

export type ScalarRecord_Substitution = {
	readonly value: Range;
	readonly value_Null: Range | null;
	readonly value_Opt?: Range;
	readonly value_Opt_Null?: Range | null;
	readonly list: Range[];
	readonly list_Null: Range[] | null;
	readonly list_Opt?: Range[];
	readonly list_Opt_Null?: Range[] | null;
};

export function isScalarRecord_Substitution(value: unknown): value is ScalarRecord_Substitution {
	return isRecord(value) &&
		checkProp(value, 'value', isRange) &&
		(checkProp(value, 'value_Null', isNull) || checkProp(value, 'value_Null', isRange)) &&
		checkOptProp(value, 'value_Opt', isRange) &&
		(checkOptProp(value, 'value_Opt_Null', isNull) || checkOptProp(value, 'value_Opt_Null', isRange)) &&
		checkProp(value, 'list', createTypedArrayGuard(isRange)) &&
		(checkProp(value, 'list_Null', isNull) || checkProp(value, 'list_Null', createTypedArrayGuard(isRange))) &&
		checkOptProp(value, 'list_Opt', createTypedArrayGuard(isRange)) &&
		(checkOptProp(value, 'list_Opt_Null', isNull) || checkOptProp(value, 'list_Opt_Null', createTypedArrayGuard(isRange)));
}

export function ScalarRecord_SubstitutionFromJSON($value: Record<string, unknown>): ScalarRecord_Substitution {
	const value = propMappedValue('value', $value, isString, RangeFromJSON);
	const value_Null = propMappedValue('value_Null', $value, isString, RangeFromJSON, 'null');
	const value_Opt = propMappedValue('value_Opt', $value, isString, RangeFromJSON, 'optional');
	const value_Opt_Null = propMappedValue('value_Opt_Null', $value, isString, RangeFromJSON, 'optional_null');
	const list = propMappedListValue('list', $value, isString, RangeFromJSON);
	const list_Null = propMappedListValue('list_Null', $value, isString, RangeFromJSON, 'null');
	const list_Opt = propMappedListValue('list_Opt', $value, isString, RangeFromJSON, 'optional');
	const list_Opt_Null = propMappedListValue('list_Opt_Null', $value, isString, RangeFromJSON, 'optional_null');
	return {
		value,
		value_Null,
		value_Opt,
		value_Opt_Null,
		list,
		list_Null,
		list_Opt,
		list_Opt_Null,
	};
}

export function ScalarRecord_SubstitutionToJSON($value: ScalarRecord_Substitution): Record<string, unknown> {
	const value = RangeToJSON($value.value);
	const value_Null = isNull($value.value_Null) ? null : RangeToJSON($value.value_Null);
	const value_Opt = isUndefined($value.value_Opt) ? undefined : RangeToJSON($value.value_Opt);
	const value_Opt_Null = isUndefined($value.value_Opt_Null) || isNull($value.value_Opt_Null) ? $value.value_Opt_Null : RangeToJSON($value.value_Opt_Null);
	const list = $value.list.map(RangeToJSON);
	const list_Null = isNull($value.list_Null) ? null : $value.list_Null.map(RangeToJSON);
	const list_Opt = isUndefined($value.list_Opt) ? undefined : $value.list_Opt.map(RangeToJSON);
	const list_Opt_Null = isUndefined($value.list_Opt_Null) || isNull($value.list_Opt_Null) ? $value.list_Opt_Null : $value.list_Opt_Null.map(RangeToJSON);

	return {
		value,
		value_Null,
		value_Opt,
		value_Opt_Null,
		list,
		list_Null,
		list_Opt,
		list_Opt_Null,
	};
}